+++ docs/guides/changelog.md
@@ -9,6 +9,7 @@ description: |-
 
 ## Unreleased
 
+- Add ICAM scale monitoring configuration to `nxos_system` resource and data source
 - Add terminal console exec timeout configuration to `nxos_system` resource and data source
 - Add terminal VTY exec timeout configuration to `nxos_system` resource and data source
 - Add terminal VTY session limit configuration to `nxos_system` resource and data source